exclude CLI output from windows tests
The windows cli is named chainlink.exe, as opposed to just chainlink.
This is represented automatically in the documentation, thanks to the
library we're using. We have a test to make sure that doesn't change
unexpectedly, but it fails on Windows because of the name change. It's a
low value test, so we just exclude it from Windows builds.