Chainlink SGX adapter stubs!
This PR adds support for a build of chainlink that includes an
alternative HTTP adapter that executes from within an enclave. This SGX
HTTP adapter is currently just a stub.

Build tags are used to disable SGX code when it is undesirable, since it
needs the SGX SDK installed and can only really be built on Linux.

Other details:
  1) SGX is disabled by default across the board
  2) SGX mode is built in simulation mode by default
  2) Integration tests do not use SGX
  3) Includes a docker compose setup for running locally