Initial implementation of doing all SGX calls on a single thread
This forces all SGX calls through a single thread that Rust manages so
that we can keep the SgxEnclave context local to a thread.

A crossbeam channel is used to send the arguments over to the worker
thread, and results are sent back via a corresponding channel.

All Go -> Enclave calls now go through one entry point with an adapter
"type" used to decide which adapter call to make.