SPIKE on using postgres as a backing store
I chose a test that uses some core models, then added a `usePG` flag to
the store, which enables postgres ORM mode.

Run the test as below:

go test -v ./services/ -run TestOneTime_RunJobAt_RunTwice

Changing usePG to false should also still pass.

This is a proof of concept for some of the changes that would need to be
made to adapt postgres under the hood, while still supporting boltDB.