Off-chain verifiable-random-function proof generation & verification
Addresses tracker story #163400928

Also, generation of RSA private key with safe-prime factors and public exponent
of 3, which is especially cheap to exponentiate by on-chain.

That this works end-to-end (off-chain) is verified in TestVRF_Generate.

The VRF.sol contract, and tests VRF_test.js, which are referenced by this
code, will be provided in a PR for story #163399542.