Move presenter logic out of FindTxByAttempt
FindTxByAttempt does this weird thing where it finds a transaction
attempt, then puts all of its details onto the transaction that it
reeturns. This makes me pretty uncomfortable because it totally changes
the semantics of the Tx object for purely presentation reasons. So I
moved this logic into the presenter.