Optimize CI test for operator_ui + fix failing test
We were previously clearing cache on every CI run, and limiting the
number of workers to 1, significantly slowing down CI times when it
came to testing operator_ui. This removes the unneeded cache removal and
increases the worker count to THREAD_COUNT/2.

The failing test that was fixed is a symptom of the --coverage flag for
jest causing it to fail at transforming class properties of
operator_ui/src/PrivateRoute. All that was done was the removal of the
class property in the favor of a manually bound class method.