Improve behavior of test lifecycle hook failures
For backwards compatibility with jasmine, failures in test lifecycle
hooks do not cause the dependent tests to fail. This causes a lot of
output noise when a hook fails since the error output will be a
combination of the hook failure and the test failures themselves.

To fix this issue, the test runner has been changed from jest-jasmine
to jest-circus. Jest-circus is the upcoming test runner that jest
will adopt in their upcoming 5.0 release. One of the benefits of
jest-circus is that lifecycle test hook failures cause the dependent
tests to fail too, shortening test failure times and reducing test
failure noise.