Swap out truffle compile for @0x/sol-compiler
The swap out of the compiler allows us to create a much more efficient
build process for developers. We gain the ability to do dockerized
compilation to gain an order of magnitude of speedup for compiling
solidity files; the ability to enable watch mode which leads to hot-
reloadable solidity tests written in typescript; and the eventual
removal of the heavyweight truffle dependency.

Some bigger changes that were made along with the swap for
compatibility and cleanup reasons:

- Generating a index barrel file for generated contract factories via
the evm/scripts/export-generated-contract-factories.ts script.

This lets us export our contract factories that were generated from the
typechain package. Allowing package consumers to access these generated
factories, giving them typesafe interaction and deployment of our
solidity contracts.

- Moving typechain generated files from dist/generated to src/generated

This is a change for future watch mode compilation and testing of
contracts, the directory src/generated is currently ignored for now,
incase we want to move it again before implementation starts.

- Move compiled solidity artifacts from build/contracts to
dist/artifacts

This was done to aggregate our compilation outputs into dist/ since we
now include javascript output, not just solidity output.