Add ethereum prefix to SA signed messages
Previously, we did not sign the SA id in the core codebase with an ethereum
message prefix added to it. This allows a potential vulnerability on the
solidity side where an oracle can be spoofed to sign a valid ethereum
transaction. Here, we add this functionality in so that all signed message data
possible by the oracle contract can not be a valid ethereum transaction