Push validateMinimumConfirmations into Execute loop
validateMinimumConfirmations checks if the run is on an uncled chain,
and is called once per task during a run. We only need to do this once
per run. This should reduce overall ethereum calls.

This also makes RunManager.Create a little bit simpler.

TestIntegration_HttpRequestWithHeaders needs to have the run's
ObservedHeight updated to satisfy the confirmations being checked before
running a task.