Add Auth Token to user
Update the user model to include fields to support token authentication.
Include the supporting web logic to support creating, deleting, removing
user token authentication.

Actions for modiying token authentication require password use.  This is
because the token authentication involves a subset of actions a user
wants to implement -for automation-.