Make external initiator specs forwar compat.
In order to improve forward compatibility, make the JSON request for
creating external initiators generic.  To do this move the name into the
params object, and make the 'body' to be posted to the external
initiator a subfield.

So
```
  {
    "type": "external",
    "name": "foo",
    "params": {
      "abc": 123
    }
  }
```

becomes
```
  {
    "type": "external",
    "params": {
      "name": "foo",
      "body": {
        "abc": 123,
      }
    }
  }
```

In order to avoid writing a migration, the database name for
Initiator.Params has been preserved.