Fix external initiator deletion
Hard delete external initiators from the database by removing the
'DeleteAt' field.  Remove also the redudant ID parameter as the Name
field functions as the entities unique ID.

The Gorm package uses the 'DeleteAt' field to hint a record was deleted,
but this conflicts with the databases UNIQUE constraint on the name
field.