Add monitoring affordance for HashToCurve iterations
The input to HashToCurve could be adversarially chosen to require an arbitrary
number of iterations, given infinite compute power. This allows the caller to
specify a function which will be called with each x ordinate considered during
the HashToCurve search, in order to count the number of iterations, or possibly
to compute offchain the expensive square roots for the y-ordinates.