XXX: Remove this commit before merging
https://github.com/smartcontractkit/chainlink-private/pull/17#discussion_r366448801

felder-cl yesterday

The tests would be more reliable if they do rely on data specifically defined
within the directory for testing. Alternatives are declaring them directly in
the file or using a 'testdata' directory.

coventry 20 hours ago Author

At least for now, this test is quite useful as-is, because it ensures that the
example key is up to date, so if it fails, it conveys important information.

felder-cl 6 hours ago

True but testing the validity of the example key should be in its own test if
its intended. It is best to keep the problem domains isolated in my opinion.
These things can accumulate and cause friction when changing one part of the
code base causes other parts to unexpectedly break.

coventry 1 minute ago Author

I'll remove it after merging PR#16 (the keystore proper) and before merging this
one. As long as the encryption/decryption algorithm is in flux, this test is
playing an important role for me.