Allow JSONParse to accept gjson (e.g. from bridge)
Rationale:

It is not at all clear to our users (or to me) what the difference is
between Copy and JSONParse and when is appropriate to use which.

I propose unifying everything under JSONParse and soft-deprecating Copy.

This PR allows JSONParse to accept gjson as well as string inputs,
making it suitable for use in parsing results from bridges (external
adapters) and making Copy (and copyPath) redundant.