Disable snapshotting optimization when debugging
sol-trace internally uses the same snapshotting technique we use, but
for the purpose of capturing traces. This conflicts with our
snapshotting technique, as ganache deletes all snapshots created after
the one you want to revert.

Because there is no synchronization mechanism between sol-trace and
our own snapshotting, any desync causes invalid snapshots to attempt
to be reverted. A discarded snapshot passed to ganache causes it to
freeze indefinitely, which creates a test timeout.