Flux Monitor: trigger by idle time
Allow chainlink operators to configure the Flux Monitor to trigger a Job
Run if it has exceeded a specified 'idle time'.  The 'idle time' is the
time for which no Job Run has been started.

The behavior is disabled if no IdleTime is defined.

https://www.pivotaltracker.com/story/show/169366221

To make testing easier, ensure the DeviationChecker has stopped
completely before returning.