Truncate the contract metatdata in the solidity binary objects
The metadata at the end of binary objects output by modern solc, varies
depending on incidental compilation context like absolute paths to source files.
https://solidity.readthedocs.io/en/v0.6.2/metadata.html#encoding-of-the-metadata-hash-in-the-bytecode

Since this suffix varies so much, it can't be included in a reliable check
that the golang wrapper is up-to-date.

This removes that metadata, for the purposes of generating a record used to
check that the golang wrapper is up-to-date.