Make construction of JSON from request log less ugly
JSON.Add extracts a map from the JSON.Result, adds the k/v pair to it, marshals
it to a string, then re-parses it to a gjson.Result.

With the number of k/v pairs I was adding, this was getting pretty ugly.

This commit means the roundtrip through JSON only happens once.