Reject RandomnessRequest logs if they come from the wrong address
The "randomnesslog" initiator now requires an address field, and only logs from
that address are watched for. That should be the address of the VRFCoordinator
contract.

This guarantees that the information in the logs the node processes has been
validated by the VRFCoordinator, so that, for instance, someone who has advance
knowledge of a prospective input seed can't trick the node into revealing the
output.