EthMock defaults to strict
Previously we were allowing calls to non-existent mocks, but logging
an error when it occured. This resulted in a log of log spew when
running tests with verbose mode enabled.

This forces all uses of ethmock to be strict by default, unless the `cltest.Lenient` flag is explicitly passed.

This cleans up a ton of useless errors from logs when tests are run with `-v` and forces new uses to be explicit about either mocking or not.

Fixes: https://www.pivotaltracker.com/n/projects/2129823/stories/170925095