Make inputs to vrf.GenerateProof uint256s instead of big.Ints
This will help get rid of the varying-time comparison on the secret key, since
now we know the length of it.

This also removes the check that the secret key is less than the secp256k1 group
order. Allowing a secret key larger than the group order only reduces the
entropy in the key infinitesimally, even if the user uniformly samples a uint256
for the key.

This addresses story #171520957.