FluxMonitor: Rename eth log channel
Attempt to improve clarity by renaming the log/error tuple to
LogOrError.  Maybe is used in other programming languages to represent a
possible non-value (Nothing).

Add a method to the type to encode retrieving the value is an either-or
action.

Also try improving its comment, and the name of the log its on.