Run each migration in its own transaction
The default way to run migrations is to run all pending migrations in
a single transaction. This is not ideal b/c if you need to run a
concurrent index, you have to run it in its own queryRunner,
which means it will execute before the main transaction COMMITs, resulting
in an error. The option 'each' runs each migration in a transaction, COMMITing
between each migration.