Wait for MinOutgoingConfirmations before proceeding in ethtx adapter
This forces the ethtx adapter to wait for
`MinRequiredOutgoingConfirmations` before marking a transaction as
confirmed and moving on.

It faithfully replicates the legacy ethtx behaviour with one exception:

`MIN_OUTGOING_CONFIRMATIONS` is now a default, and
`minRequiredOutgoingConfirmations` may now also be set on the ethtx task
spec to optionally override this default.