Process higher value eth_tx at the end
The purpose of this is to send as many transactions as possible before
we hit an "out of Eth" error.

If somehow a transaction has been created with a ridiculously large
value that exceeds the funds in the wallet, at least this way it won't
block the queue and the node can limp on until the operator manually
fixes it.