Deflake TestFluxMonitorAntiSpamLogic
This adds a missing .Commit() call to the simulated-backend ethereum client,
which was causing flux-monitor answers to be submitted in an unexpected order.

It also tightens up synchronization of the test with the simulated backend a
little, by waiting for log emmissions.

I've run this a few hundred times without a failure.