Autoset P2P_PEER_ID on start if there is only one key
The application now works as follows:

0 p2p keys in DB
================

Application will autocreate a key on startup and use this as the
default.

1 p2p key in DB
================

`P2P_PEER_ID` MAY be specified.

If specified, it must match the existing key.

2 or more p2p keys in DB
========================

`P2P_PEER_ID` MUST be set and MUST match one of the keys in the DB if
off-chain reporting is enabled, otherwise the application will fail to
start.