Use retry loop instead of waiting for advisory lock (#3835)
* Use retry loop instead of waiting for advisory lock

Waiting for an advisory lock for a really long time (hours/days/weeks) is problematic.

This is because while waiting for a lock, postgres must hold a
transaction open. This causes problems because postgres must hold the
MVCC snapshot from the start of transaction time and prevents automatic
vacuum and other garbage collection operations.

A common use-case is to run multiple chainlink nodes, one of which is
the "primary" and the other ones wait for the lock. This use case is
causing bad postgres performance because of the long running
transactions that are held open.

This commit replaces the wait with a periodic retry, which does not hold
open a transaction.

* Set some sanity limits on lock/tx timeouts

* Add note to changelog

* Do not continue to try and run queries if the lock timeout fires

* Remove useless var

* Fix errcheck

* Cannot set timeouts at session level due to connection pooling

* Set some sane values for production database connections

* Convert maxOpenConns/maxIdleConns to env vars

* Fix vet

* PR nit