Replace all metadata hashes with a constant value
It seems that if the `new` solidity keyword is used by a contract, solc will
generate a binary which includes the entire binary of the contract `new` is
deploying, including an extra metadata hash.

This commit replaces all matches for the metadata hash in the binary.

This is perhaps slightly fragile; it's conceivable that a metadata-hash match
could happen accidentally. There's about 128 bits of difficulty in coming up
with such a hash (16 fixed bytes, and ignoring the fixed distances required
between the matches, and the fact that there are multiple candidates in a given
binary..)