Prevent rare foreign key error "log_broadcasts_eth_log_id_fkey"
In some very rare cases if we delete an eth_log whilst simultaneously
upserting a log broadcast we can insert stale data (an eth_log id that
no longer exists) causing foreign key errors.

In this case we use optimistic locking such that if we receive this
foreign key violation we act as if the log was inserted successfully
then immediately deleted again.