Retry polling when an idle timer fails
When an idle timer fails to submit, a retry ticker is started which attempts to poll with a backoff. The backoff period is defaulted to a minimum of 1 minute up
to a maximum of 1 hour. If the idle timer period is set below these values, it
will tick at the idle timer period duration.