Update fastgen.py to new contract situation
fastgen.py is a script for rapdidly generating golang wrappers of solidity
contracts during develop, to speed the compile/test/debug cycle.

This commit skips over the OffchainAggregator wrapper, since that is built in
its own way, and the contract is developed outside chainlink core.

It also makes the wrapper package directory, if necessary, which avoids an
unnecessary failure mode.