OCR keys now use a presenter in the API
This commit changes the API to render an OCRKey presenter instead of the OCRKey model. This allows us to remove presentation specific code and tags from the models.

In addition we allow the presenter to define how it renders itself into a table by implementing a new `TableRenderer` interface. This allows us to begin removing the large switch statement in `renderer.go`