Re-enable transactional migrations
Disabling this means that if a migration fails for any reason, the
database is left in an undefined state.

Unfortunately gorm's behaviour on this is bad since ALL migrations are
wrapped in the transaction, so it is not possible to disable it on a
per-migration basis.

For now, we should re-enable it pending this card: https://app.clubhouse.io/chainlinklabs/story/7191/write-our-own-migration-helper-to-replace-gorm-s-one