Fix flaky test Test_UpkeepExecuter_PerformsUpkeep_Happy; make keeper inserts atomic
The test was flaking occasionally because keeper was inserting the eth_tx then
updating the last run height.

In this case, if you get unlucky with the timing and check eth_tx right
after it has been inserted (but before the run height was updated) you
would read the run count at the wrong time and the test would fail.

Fixed by changing the order of the code and test so that it checks for
run height only after the run has been inserted, guaranteeing the order.

In addition, create eth transaction and setting the last run height have
been transactionalised, making them atomic. This way either run height
gets updated AND eth_tx is sent, or neither happen. It is no longer
possible to send the eth_tx and then fail to update the run height.