Adds an endpoint to cancel a job via the Feeds Manager (#5041)
* Adds an endpoint to cancel a job via the Feeds Manager

This commits adds an endpoint to cancel a job that was started by the Feeds Manager.
Cancelling a job will update the proposal record status to cancelled, deletes
the running job, and notifies the Feeds Manager. In the event of an error, the transaction
is rolled back and data state is preserved. Due to the nature of how jobs are run, the
running job is stopped but will be restarted again once the event loop of the spawner runs.

A job which was started by the Feeds Manager can only be stopped in the Feeds Manager UI.
The endpoint to delete a job now has an additional check to ensure that the job was not
started by the Feeds Manager.