Fix regression where gas estimator would not use current gas price
Accidentally introduced a regression in
9303d55162297c1124a637d121b173244a7d19e8 with the refactor of gas
estimator to support Optimism.

We used to bump on top of the latest price if that was larger. This was
inadvertatently broken so that bumping no longer uses current gas price
from the block history estimator if it's larger.

This commit reintroduces logic such that if the current price is larger
than the bumped price, the current price will be used instead. This
should help make Chainlink more responsive and save a big of gas under
very steep spikes.