Fix regression where gas estimator would not use current gas price (#5113)
Accidentally introduced a regression in
c6b9186b8009c986bca39a925030d69c7e9c1a8e with the refactor of gas
estimator to support Optimism.

We used to bump on top of the latest price if that was larger. This was
inadvertatently broken so that bumping no longer uses current gas price
from the block history estimator if it's larger.

This commit reintroduces logic such that if the current price is larger
than the bumped price, the current price will be used instead. This
should help make Chainlink more responsive and save a big of gas under
very steep spikes.