Hard panic in task runs; scheduler timeout for pipeline run
Recovering panics in the pipeline can lead to inconsistent state and
hang the pipeline indefinitely. It is better to hard panic immediately
and allow the node to be cleanly restarted.

In addition, the pipeline runner should have some kind of sanity
timeout. This commit re-adds a deadline for JobPipelineMaxRunDuration
which appears to have been removed by accident at some point.

---

Completely remove WrapRecover; hard panic everywhere

Recover-ing gives illusion of safety but in reality it makes things more
dangerous.

In almost all cases its safer to hard panic the node than try to recover
and keep going since panicking can leave things in an unknown state.

Node operators typically run with auto-restart enabled, if the node
panics it will be booted again from a clean state, and we immediately
get a clear stacktrace of where the problem is.

We have seen bugs that would have been uncovered much sooner and caused
fewer problems if we had not suppressed the panic.

There is one and only one case I can think of where recovering panics is
acceptable, and that is in database transactions where it can be
intercepted in order to rollback the transaction, then re-panicked
again - and even that must be done very carefully with an additional
panic added with a timeout in case the rollback hangs.

Closes: #5384
Approved by: samsondav