Handle errors gracefully
In the unlikely event we fail to hash the vrf public key or get the
uncompressed version, we simply return an error in place of the field.
It is highly unlikely for such errors to occur (if they do, the app as a
whole will probably be failing all over the place) but it's good
practice to handle the errors anyway instead of completely ignoring
them.