Remove P2PPeerID from OCR spec
When peerwrapper was initially written, it was assumed that support for
multiple P2P Peer IDs would be necessary in short order.

It turned out that this requirement was much lower priority than
expected, so that support was never needed.

Keeping P2PPeerID on the OCR spec is therefore unnecessarily complicated
and doesn't do anything useful since only one peer is ever actually
supported.

This commit removes support for setting P2PPeerID on an individual spec.