Do not block main sending thread on sendonly nodes
Fiews has reported when using a slow/broken node for secondary it causes
reverts, due to holding up the main thread while waiting for a response.

Since sendonlys are always best effort, it doesn't make sense to block
the main thread on these and instead to "fire-and-forget", returning as
soon as the primary has returned.

Also reduce eth timeout from 15s->10s.