fix backend to respond to input pipeline window
Entries placed in the input pipeline window are ignored
because the REST message to the backend uses run id,
which then calls a run function which ignore messages
sent to the backend.

This patch changes the backend so that if a integer
run id is called, that it will convert it to a external
job id and then run that.