Fixes "Killed: 9" code-signing crash on MacOS
Apple does not support using "cp" to update a signed binary with a newer
version of it.

Their policy is that once a binary is signed, its contents is never expected
to change or the OS reserves the right to immediately kill it as soon as
anything tries to run it in case it's malware.

The only supported means of replacing a signed binary on recent versions of
MacOS is to completely remove it first and create a brand new file (which must
have a different inode) with a new signature:
https://developer.apple.com/documentation/security/updating_mac_software/