Rework audit logger system as standalone interface and struct instead of nesting and working into the current Logger system.
This is a better implementation because it does not polite the logger interface (adding .Audit that only  this implementation uses), and does not require special wrapping

Previously the audit logger would just forward through all other function calls. Now the auditLogger is a sibling field adjacent to the logger field in the Application struct. It is parameterized and passed through the call stack.

Update call sites to pass context for threaded calls and move auditLogger to its own package