Retry indefinitely on TxFeeExceedsCapError (#7019)
In the case of multiple RPC nodes, some can have a configured cap whilst
others do not. Because we broadcast in parallel to all nodes, this can
lead to a TxFeeExceedsCapError even if the tx has been accepted into the
mempool by another node.

Marking fatal error could lead to nonce re-use, so the correct thing to
do in this case instead is to continue retrying indefinitely and move
ahead only once a tx with that nonce has been accepted.