Add support for besu's non-unique logIndex
The besu client doesn't strictly conform to the eth jsonrpc spec.
Instead of logIndex being unique within a block, it's only unique
within a transaction.  This adds TxIndex to the key we use to
track duplicate logs, making the chainlink node robust enough to
support besu or any other clients which behave this way.

Also:
 - Update vrf's custom deduper with same change
 - Update tests