Fix failing short test with DB requirement
All tests in `TestETHCallTask` require a DB instance to run as
indicated by `db := pgtest.NewSqlxDB(t)`. The test should be skipped
as early as possible due to the DB dependency. The db initialization
line was moved to the top of the test setup to skip early for short
tests.