Introduce unit tests for RegistrySynchronizer
To verify the log output is formatted as expected, a unit test was added to
`RegistrySynchronizer.syncUpkeepWithCallback`. To accomplish a unit test on
this private member function, the test needed to be in the keeper package
directly which eliminated the ability to use `cltest` for testing.

To avoid the eth client dependency, I introduced a new interface that private
functions can receive as a parameter instead of a `RegistryWrapper` directly.
This allows function isolation with minimal test setup and simplifies unit
testing without the need to rely on the cltest package.