UpkeepID as a formal type
UpkeepIdentifier was created as a wrapper for `utils.Big` to provide support for
specific formatting associated with the underlying data. The future intention of
this new type is to use it in place of `utils.Big` where the data should be
considered an upkeepID. This will allow type checking to apply to instances where
an upkeepID is the intended data to be passed through.

In addition, errors from the registry wrapper were refactored to provide more
context to the errors and leverage the value of error wrapping to identify errors
as having originating from contract communication.