Pretty print upkeepID as prettyID in job spec
The same pretty print upkeepID function applies to the prettyID field
of the job spec. The upkeepID was kept the same; an int represented
as a string.

To better ensure correctness of the entire job spec, including the
upkeepID, UpkeepExecuter.execute was refactored by pulling the job
spec build step into its own function. That function is now covered
by a unit test.