fix: MIN_INCOMING_CONFIRMATIONS does not override job spec value (#7385)
This has caused a lot of confusion for operating VRF nodes. Since
`minIncomingConfirmations` is a required field on the VRF job spec,
it will always be provided and set to a reasonable value (provided
the operator knows what they're doing).

Overriding conditionally based on which value is greater has caused
more of an operational burden, and it's just easier to have a single
source of truth on the number of confirmations, and the job spec is the
ideal source of truth for this information.