Remove redundant/inaccurate hint in OCR / OCR2 error messages (#7473)
* Remove redundant/inaccurate hint in OCR2 error message

The peer wrapper will report exactly why it failed to start.  No need
to guess why later when OCR2 is trying to start.

In the case of OCR, it may or may not be the right guess.  In the case
of OCR2 it will almost always be the wrong guess.  (The peer wrapper
might fail to start if it's configured to use v1, but that has nothing
to do with OCR2.  If OCR2 is enabled, it's more likely due to one of
the v2 env vars not being set, since that's the only protocol it uses.)

* Remove punctuation