Update logic for de-duping contract addresses
Previous logic had several issues:

- When a new job is created with no chain id specified (chain_id=NULL), we
were not checking to see if there were any existing jobs with an explicit
chain id specified.  We were checking for existing jobs with evm_chain_id=NULL,
but generating a nil pointer dereference in that case instead of the intended error
message.

- When a new job is created with a chain id specified, we were checking
for jobs with the same chain id.  But only checking for jobs with
evm_chain_id=NULL in the case where the new job's chain id matches the
current default chain id.

Examples of loopholes in current logic:

- Existing evm_chain_id=1, default chain id 1, new job added with chain_id=NULL
   This was allowing the job to be added, immediately causing both to run on chain id 1

- Existing evm_chain_id=NULL, default chain id 1, new job added with chain_id=2
   This was allowing the job to be added, which would be okay as long as the
   default chain id never changes.  But if it's later changed to 2, then both
   will be running on the same chain with the same contract address.

New logic is simpler & tighter:

- Never add a new job with the same contract addr unless both new and
  existing jobs have explicit non-NULL chain ids and the chain ids are different.

- If chain id is missing from job spec, explicitly fill in the default
  chain id, rather than writing the NULL chain id to the db

We will still allow existing chain ids to be NULL in the db, and these will run
  on any chain that the current default is set to.  But this will not be supported
  for new jobsd.  Going forward, if you add a job with no chain id, it
  will only run on the current default chain, not on any chain you may change
  the default to in the future.  This forces the user to take the safer
  approach of adding a second job if they want a job to be able to run on more
  than one chain.