Simplifies FMS configuration initialization
With the update to multi chain support, FMS no longer needs the global
OCR configuration values as part of the config interface. The OCR1
validation function was changed to use the configuration set in the
chain sets, rather than env vars.

This change removes the unused config interface declarations, and now
allows us to pass in the general config object when initializing Feeds
Manager.