Fixes job replacement for legacy jobs with no evm chain id association
For very old jobs, the `evm_chain_id` column is set to `NULL` which causes
`FindJobByAddress` to fail because it requires both the contract address
and evm chain id to match.

Any new jobs since the introduction of the `evm_chain_id` column,
automatically inserts this value from either the job spec `EVMChainID`
field or the default chain id value set in the env var.

This change updates the query to allow for a match on a NULL value in
the `evm_chain_id` column. Since there is a partial unique constraint
on a contract address with a NULL evm chain id, we can safely add this
OR condition to the orm method and be sure it will only return a single
record.