Fix lock timeout CI failures
Tracked down the root cause to parallel tests in this particular package
that cause many concurrent executions of NewSession. In turn, this
tries to do an insert to the DB, which is where the lock contention
is occurring.

Change the test code to override the database lock timeout based on the
test deadline, when applicable. When there is no test timeout, increase the
lock timeout by 4x, which is an educated guess based on CI failures