Cancel the existing FMS managed job when a new one is approved (#8525)
This is a quality of life improvement for node operators when approving
job proposals with FMS. Previously any updates to a job proposal would
require the user to manually cancel the existing spec, before approving
the new one. With this change, approving a new job will automatically
cancel the existing approved job before running the new job.