BCF-2099 refine nurse test and logging
The test was flaky because of timing with file creation. Change test
to check for eventually existence of profiles. This is necessary due
to concurrent nature of the profile gathering.

Change test to register a check function, like normal usage, rather
than directly call GatherVitals

Add logging to the gather funcs for debugging.