core/chains/evm/log: replay set long query timeout
There are cases where a large amount of logs need to be marked as
unconsumed, particularly for VRF replays that occur a while back. Using
a longer query timeout in those cases would make sense because the
alternative is deleting rows from the table (which requires infra
assistance) in the event the MarkBroadcastsUnconsumed query times out.