Return the job proposal name in the GQL API (#8555)
* Return the job proposal name in the GQL API

Users want to be able to fetch the name of the job proposal, however
job proposals themselves don't have a name, but rather the job proposal
spec has the name in it's TOML spec definition.

This change adds a new `name` column to the job proposal table, which
is populated by parsing the name out of the TOML spec definition any
time a job is proposed. The GQL layer has been updated to return this
new field.

Existing proposals' names are updated during the migration by performing
a regex capture on the latest spec definition for a proposal, and
inserting the value onto the proposals table. If the regex fails to
match, the name is not inserted for that proposal.