Fix flaky logpoller test, add t.Parallel() to tests
There were two racy issues with this test seen in CI

One was a -race failure seen once in CI.  Not reproducible locally, but was possible because
lp.ctx was intialized once, then reinitialized to the same thing in Start(), and read somewhere
in between or afterwards.  Arguably not an actually a race since the value doesn't change, but
the write -> read -> write pattern must have triggered the race detector.

The other was never detected as a race, but occasionally there were 5
log messages returned instead of 4.  The extra log message is
irrelevant, so test of exact number has been removed.

Also, replaced Sleep() with Eventually() to be more robust in case there are some unusually large
scheduling delays.  Haven't seen this happen yet but being pro-active.