BCF-2117 More thorough block JSON testing
This is a warmup for BCF-991 faster JSON unmarshal'ing

- Adds tests for Transaction and TxType JSON un/marshal'ing
- Adds the baseline benchmark

The benchmark needs to marshal and unmarshal JSON in order to easily
create various size input for the number of transactions. The roundtrip
did not work because we had custom UnmarshalJSON logic, but nothing for
Marshaling.

Therefore, in order to write the benchmark efficiently, I had to implement
some minimal JSON marshal'ing logic.
This is tested as part of the new unit tests.

Once this is merged, we can fix BCF-991 and be confidient about the change