BCF-1835: rm legacy vars from errors and logs
We used the legacy.env to search the code base for references

This change updates the reference to the new TOML name in all errors and logs

We choose the fully qualified TOML name ~ x.y.z because this was the
easiest form for search the TOML config files in core/config/v2/docs, and
therefore it is expected that it provides the best UX.

The scope here is limited to errors and logs. Some references to legacy
var remain. Specifically we excluded the following patterns in while replacing:

handler.go, **integration**, *scripts/**, *migrations/**